backend/requirements.txt
asgiref==3.5.2
bcrypt==4.0.0
black==22.8.0
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_83242e7e-f82d-4a71-8ef2-9d71d212d249gu_wxmeq/croots/recipe/certifi_1655968827803/work/certifi
click==8.1.3
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mypy-extensions==0.4.3
mysqlclient==2.1.1
pathspec==0.10.1
platformdirs==2.5.2
PyJWT==2.4.0
python-dotenv==0.21.0
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
